Move the dhcp pkg under internal:
This minimizes our public/external API surface.
