+++ backend/file/file_test.go
@@ -19,7 +19,7 @@ import (
 	"github.com/go-logr/stdr"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/tinkerbell/smee/dhcp/data"
+	"github.com/tinkerbell/smee/internal/dhcp/data"
 )
 
 func TestNewWatcher(t *testing.T) {
