+++ cmd/smee/backend.go
@@ -6,7 +6,7 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/tinkerbell/smee/backend/file"
 	"github.com/tinkerbell/smee/backend/kube"
-	"github.com/tinkerbell/smee/dhcp/handler"
+	"github.com/tinkerbell/smee/internal/dhcp/handler"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
