+++ cmd/smee/main.go
@@ -22,9 +22,9 @@ import (
 	"github.com/insomniacslk/dhcp/dhcpv4/server4"
 	"github.com/tinkerbell/ipxedust"
 	"github.com/tinkerbell/ipxedust/ihttp"
-	"github.com/tinkerbell/smee/dhcp"
-	"github.com/tinkerbell/smee/dhcp/handler"
-	"github.com/tinkerbell/smee/dhcp/handler/reservation"
+	"github.com/tinkerbell/smee/internal/dhcp"
+	"github.com/tinkerbell/smee/internal/dhcp/handler"
+	"github.com/tinkerbell/smee/internal/dhcp/handler/reservation"
 	"github.com/tinkerbell/smee/ipxe/http"
 	"github.com/tinkerbell/smee/ipxe/script"
 	"github.com/tinkerbell/smee/metrics"
