--- dhcp/example/kubeBackend/hardware.yaml
@@ -1,38 +1,0 @@
-apiVersion: "tinkerbell.org/v1alpha1"
-kind: Hardware
-metadata:
-  name: sm01
-  namespace: default
-spec:
-  disks:
-    - device: /dev/nvme0n1
-  metadata:
-    facility:
-      facility_code: onprem
-    manufacturer:
-      slug: supermicro
-    instance:
-      userdata: ""
-      hostname: "sm01"
-      id: "de:ad:c0:de:ca:fe"
-      operating_system:
-        distro: "ubuntu"
-        os_slug: "ubuntu_20_04"
-        version: "20.04"
-  interfaces:
-    - dhcp:
-        arch: x86_64
-        hostname: sm01
-        ip:
-          address: 192.168.2.17
-          gateway: 192.169.2.1
-          netmask: 255.255.255.0
-        lease_time: 86400
-        mac: de:ad:c0:de:ca:fe
-        name_servers:
-          - 192.168.2.1
-          - 10.1.1.11
-        uefi: true
-      netboot:
-        allowPXE: true
-        allowWorkflow: true
