+++ go.mod
@@ -18,7 +18,6 @@ require (
 	github.com/prometheus/client_golang v1.18.0
 	github.com/tinkerbell/ipxedust v0.0.0-20231215220341-a535c5deb47a
 	github.com/tinkerbell/tink v0.9.0
-	github.com/tonglil/buflogr v1.0.1
 	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.46.1
 	go.opentelemetry.io/otel v1.21.0
 	go.opentelemetry.io/otel/trace v1.21.0
