+++ internal/dhcp/handler/reservation/handler_test.go
@@ -19,8 +19,8 @@ import (
 	"github.com/insomniacslk/dhcp/dhcpv4"
 	"github.com/insomniacslk/dhcp/iana"
 	"github.com/insomniacslk/dhcp/rfc1035label"
-	"github.com/tinkerbell/smee/dhcp/data"
-	"github.com/tinkerbell/smee/dhcp/otel"
+	"github.com/tinkerbell/smee/internal/dhcp/data"
+	"github.com/tinkerbell/smee/internal/dhcp/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"golang.org/x/net/ipv4"
 	"golang.org/x/net/nettest"
