+++ internal/dhcp/otel/otel.go
