+++ internal/dhcp/otel/otel_test.go
